refactor(audio): change mixer to float-based accumulation with headroom
- Change generateSampleForChannel return type from int16_t to float to preserve dynamic range
- Implement accumulator mixing with fixed headroom (0.25) and single final clamp
- Add peak detection logging to diagnose clipping issues
- Create optimization plans for non-linear mixing and ESP32 distortion analysis

+++ include/drivers/esp32/ESP32AudioScheduler.h
@@ -64,7 +64,7 @@ namespace pixelroot32::audio {
         void playCurrentNote();
         
         AudioChannel* findFreeChannel(WaveType type);
-        int16_t generateSampleForChannel(AudioChannel& ch);
+        float generateSampleForChannel(AudioChannel& ch);
     };
 
 } // namespace pixelroot32::audio
